// apps/api/src/generated/kysely/automations/RemoteAdapter.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ExternalServiceCredentialsId } from './ExternalServiceCredentials';
import type { ColumnType, Selectable, Insertable, Updateable } from 'kysely';

/** Identifier type for automations.remote_adapter */
export type RemoteAdapterId = string & { __brand: 'automations.remote_adapter' };

/** Represents the table automations.remote_adapter */
export default interface RemoteAdapterTable {
  id: ColumnType<RemoteAdapterId, RemoteAdapterId | undefined, RemoteAdapterId>;

  team_id: ColumnType<string, string, string>;

  adapter_type: ColumnType<string, string, string>;

  base_url: ColumnType<string, string, string>;

  auth_strategy: ColumnType<unknown, unknown, unknown>;

  credentials_id: ColumnType<ExternalServiceCredentialsId | null, ExternalServiceCredentialsId | null, ExternalServiceCredentialsId | null>;

  manifest: ColumnType<unknown, unknown, unknown>;

  created_at: ColumnType<Date, Date | string | undefined, Date | string>;

  updated_at: ColumnType<Date, Date | string | undefined, Date | string>;
}

export type RemoteAdapter = Selectable<RemoteAdapterTable>;

export type NewRemoteAdapter = Insertable<RemoteAdapterTable>;

export type RemoteAdapterUpdate = Updateable<RemoteAdapterTable>;